+++ pkg/providers/appsec/testdata/TestResURLProtectionAction/delete.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+# Empty config to test deletion