+++ pkg/providers/appsec/testdata/TestResURLProtectionAction/missing_security_policy_id.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+resource "akamai_appsec_url_protection_action" "test" {
+  config_id                 = 43253
+  url_protection_rule_id    = 135355
+  max_rate_threshold_action = "alert"
+}